Extract fleet scores invalidation tests and clarify wiring docstring.
Move deferred-notification and per-player invalidation coverage into
test_fleet_scores_invalidation.py; rename the defer test to reference
ledger notifications.
