+++ packages/api/tests/test_fleet_persistence.py
@@ -1074,145 +1074,3 @@ def test_stale_chain_anchor_skipped_during_gap_fill(persistence, load_turn, memo
     assert rematerialized_110 is not None
     assert rematerialized_110.materialization_version == FLEET_MATERIALIZATION_VERSION
     assert all(record.record_id != "stale-rec" for record in rematerialized_110.players[0].records)
-
-
-def test_gap_fill_defers_snapshot_notify_until_chain_completes(
-    persistence,
-    load_turn,
-    memory_backend,
-):
-    """Per-player put_ledger during gap-fill must not notify; flush runs after the chain."""
-    from api.analytics.turn_roster import iter_turn_players
-
-    turn_111 = load_turn(111)
-    assert turn_111 is not None
-    turn_110 = load_turn(110)
-    assert turn_110 is not None
-    _put_provenance_final_snapshot(persistence, 628580, 1, turn_110)
-
-    inference_persistence, inference_materialization = _inference_materialization_for_fleet(
-        memory_backend,
-        load_turn,
-    )
-    _seed_scores_rows_for_all_players(inference_persistence, turn_111)
-
-    callback_events: list[tuple[int, int]] = []
-    persistence.on_ledger_persisted = lambda _g, _p, turn_number, player_id: callback_events.append(
-        (turn_number, player_id)
-    )
-
-    put_ledger_calls = 0
-    original_put_ledger = persistence.put_ledger
-
-    def counting_put_ledger(*args, **kwargs):
-        nonlocal put_ledger_calls
-        put_ledger_calls += 1
-        return original_put_ledger(*args, **kwargs)
-
-    persistence.put_ledger = counting_put_ledger  # type: ignore[method-assign]
-
-    snapshot = get_or_materialize_fleet_snapshot(
-        persistence,
-        628580,
-        1,
-        turn_111,
-        load_turn=load_turn,
-        inference_materialization=inference_materialization,
-    )
-
-    assert snapshot.turn == 111
-    roster_size = len(list(iter_turn_players(turn_111)))
-    assert roster_size > 1
-    assert put_ledger_calls >= roster_size
-    assert len(callback_events) == roster_size
-    assert all(turn_number == 111 for turn_number, _ in callback_events)
-
-
-def test_gap_fill_emits_deferred_scores_invalidation_after_chain_completes(
-    persistence,
-    load_turn,
-    memory_backend,
-):
-    """After gap-fill, newly complete fleet@(T-1) invalidates scores@T (not mid-chain)."""
-    from api.analytics.turn_roster import iter_turn_players
-
-    inference_persistence, inference_materialization = _inference_materialization_for_fleet(
-        memory_backend,
-        load_turn,
-    )
-    invalidation = InferenceInvalidationService(
-        inference_persistence,
-        fleet_persistence=persistence,
-    )
-    invalidation.wire_scores_invalidation_to_fleet_persistence()
-
-    turn_112 = load_turn(112)
-    assert turn_112 is not None
-    turn_111 = load_turn(111)
-    assert turn_111 is not None
-    turn_110 = load_turn(110)
-    assert turn_110 is not None
-    _put_provenance_final_snapshot(persistence, 628580, 1, turn_110)
-    _seed_scores_rows_for_all_players(inference_persistence, turn_111)
-    _seed_scores_rows_for_all_players(inference_persistence, turn_112)
-
-    snapshot = get_or_materialize_fleet_snapshot(
-        persistence,
-        628580,
-        1,
-        turn_112,
-        load_turn=load_turn,
-        inference_materialization=inference_materialization,
-    )
-
-    assert snapshot.turn == 112
-    for player in iter_turn_players(turn_112):
-        assert inference_persistence.get_row(628580, 1, 112, player.id) is None
-
-
-def test_fleet_ledger_persisted_invalidates_scores_row_for_player_only(
-    persistence,
-    load_turn,
-    memory_backend,
-    monkeypatch,
-):
-    from api.analytics.turn_roster import iter_turn_players
-
-    inference_persistence, _ = _inference_materialization_for_fleet(memory_backend, load_turn)
-    turn_112 = load_turn(112)
-    assert turn_112 is not None
-    players = list(iter_turn_players(turn_112))
-    player_p = players[0].id
-    player_q = players[1].id
-    _seed_scores_rows_for_all_players(inference_persistence, turn_112)
-
-    rescheduled_players: list[int] = []
-    all_rescheduled: list[None] = []
-
-    def spy_reschedule_row(_scope, player_id, **_kwargs):
-        rescheduled_players.append(player_id)
-
-    def spy_reschedule_all(_scope, **_kwargs):
-        all_rescheduled.append(None)
-
-    monkeypatch.setattr(
-        "api.services.inference_invalidation_service.reschedule_inference_row",
-        spy_reschedule_row,
-    )
-    monkeypatch.setattr(
-        "api.services.inference_invalidation_service.reschedule_all_inference_rows",
-        spy_reschedule_all,
-    )
-
-    invalidation = InferenceInvalidationService(
-        inference_persistence,
-        fleet_persistence=persistence,
-    )
-    invalidation.wire_scores_invalidation_to_fleet_persistence()
-
-    invalidation.on_fleet_ledger_persisted(628580, 1, 111, player_p)
-
-    assert inference_persistence.get_row(628580, 1, 112, player_p) is None
-    assert inference_persistence.get_row(628580, 1, 112, player_q) is not None
-    assert rescheduled_players == [player_p]
-    assert all_rescheduled == []
